refactored and working pending pod enricher
